src/main/java/com/zipcodewilmington/phonebook/PhoneBook.java
package com.zipcodewilmington.phonebook;
import java.util.*;

/**
 * Created by leon on 1/23/18.
 */
public class PhoneBook {

    //instance variable
    private TreeMap<String, ArrayList<String>> treeMap;


    // void "no arguments" constructor -creates empty map
    public PhoneBook() {
        this.treeMap= new TreeMap<String, ArrayList<String>>();
    }

    // single argument constructor -type Map -creates new map w/ same key/value mappings as argument
    public PhoneBook(TreeMap treeMap) {
        this.treeMap = treeMap;
    }

    //add a name & number entry, add number
    public void inputEntry(String name, ArrayList<String> number) {
        if (!treeMap.containsKey(name)) {
            treeMap.put(name, number);
        }
    }

    //add numbers to existing entries
    //no test cases yet
    public void addNumberToEntry(String name, String number) {
        for (Map.Entry<String, ArrayList<String>> entry : treeMap.entrySet()) {
            if (name.equals(entry.getKey())) {
                ArrayList<String> value = entry.getValue();
                value.add(number);
            }
        }
    }

    //remove numbers from existing entries
    //no test cases yet
    public void removeNumberFromEntry(String name, String number) {
        for (Map.Entry<String, ArrayList<String>> entry : treeMap.entrySet()) {
            if (entry.getKey().equals(name)) {
                if (entry.getValue().contains(number)) {
                    entry.getValue().remove(number);
                }
            }
        }
    }

    //remove a name & number entry
    public void removeEntry(String name) {
            treeMap.remove(name);
    }

    //find phone number lookup by name
    public ArrayList<String> lookup(String name) {
        return treeMap.get(name);
    }

    //reverse lookup
    public String reverseLookup (String number) {
        String name = "";
        for(Map.Entry<String, ArrayList<String>> entry: treeMap.entrySet()) {
            ArrayList<String> value = entry.getValue();
            for(String revnumber: value) {
                if(number.contains(revnumber)) {
                    name = entry.getKey();
                }
            }
        }
        return name;
    }

    //print out all of the entries in PhoneBook
    public String display(){
        StringBuilder printOut = new StringBuilder();
        for(Map.Entry<String, ArrayList<String>> entry: treeMap.entrySet()) {
            printOut
                    .append(entry.getKey())
                    .append(" ")
                    .append(entry.getValue().toString())
                    .append("\n");
        }
        return printOut.toString();
    }

}
